Updated GUI functionality for moddifying current inventory as a new title and adding inventory disk SN
