+++ mainWindowV2.1.ui
@@ -11,7 +11,7 @@
    </rect>
   </property>
   <property name="windowTitle">
-   <string>MainWindow</string>
+   <string>Redbox Inventory Manager</string>
   </property>
   <widget class="QWidget" name="centralwidget">
    <layout class="QVBoxLayout" name="verticalLayout_4">
@@ -23,7 +23,7 @@
        </font>
       </property>
       <property name="currentIndex">
-       <number>0</number>
+       <number>2</number>
       </property>
       <widget class="QWidget" name="ProductsPage">
        <property name="geometry">
@@ -299,7 +299,7 @@
                </widget>
               </item>
               <item>
-               <widget class="QTextEdit" name="description">
+               <widget class="QPlainTextEdit" name="description">
                 <property name="minimumSize">
                  <size>
                   <width>0</width>
@@ -462,6 +462,9 @@
               </item>
               <item row="1" column="2">
                <widget class="QPushButton" name="addToInventory">
+                <property name="enabled">
+                 <bool>false</bool>
+                </property>
                 <property name="sizePolicy">
                  <sizepolicy hsizetype="Minimum" vsizetype="Preferred">
                   <horstretch>0</horstretch>
@@ -499,8 +502,8 @@
         <rect>
          <x>0</x>
          <y>0</y>
-         <width>356</width>
-         <height>145</height>
+         <width>1006</width>
+         <height>580</height>
         </rect>
        </property>
        <attribute name="label">
@@ -590,7 +593,6 @@
     <property name="title">
      <string>Tools</string>
     </property>
-    <addaction name="actionInsert_Disk_SN"/>
     <addaction name="actionSearch_By_Disk_SN"/>
    </widget>
    <widget class="QMenu" name="menuHelp">
